Add Flatpak build files and docs
This patch adds documentation on updating the bindings oneself and therefore closes #5 and possibly #3.

Furthermore, it adds a way to generate the bindings using Flatpak.
This comes in handy, if one wants to target Flatpak as a build target.
Using this build target, one has the newest version of Libadwaita available to them.